pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "worldtensor"
version = "0.1.0"
description = "Harmonized 0.25° global gridded dataset for Earth System Foundation Models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "xarray>=2023.1",
    "rioxarray>=0.14",
    "rasterio>=1.3",
    "netCDF4>=1.6",
    "numpy>=1.24",
    "geopandas>=0.13",
    "shapely>=2.0",
    "geocube>=0.4",
    "dask[complete]>=2023.1",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "click>=8.1",
    "cartopy>=0.21",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "cdsapi>=0.7",
    "earthaccess>=0.15",
    "requests>=2.30",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0",
    "torchgeo>=0.5",
    "scikit-learn>=1.3",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
